requirements.txt
flask
requests
google-generativeai
mistralai
huggingface_hub
python-dotenv